restore texts and more file upload method
